openssl: Add no_autoload_config option
Prior to this commit it was necessary to pass OPENSSL_CONF=/dev/null
in the environment of the running executable to prevent libcrypto.so
from loading the default openssl.cnf configuration file. With this
commit it is now also possible to disable loading the default
configuration file at build time by setting
openssl:no_autoload_config=True.

By default libcrypto.so will autoload a default configuration - on
Linux it will be /etc/ssl/openssl.cnf. This file is under normal
circumstances not associated with the conan-provided openssl library
and can contain instructions that will lead to errors. For instance
the configuration file can contain instructions to load additional
providers which can lead to errors like these:

    error:25066067:DSO support routines:dlfcn_load:could not load the shared library
    error:25070067:DSO support routines:DSO_load:could not load the shared library
    error:0E07506E:configuration file routines:module_load_dso:error loading dso
    error:0E076071:configuration file routines:module_run:unknown module name