openssl: Enable valid block of options for openssl >= 1.1.0
The options zlib, rc4, rc5, and md2 have been deleted since commit
fdc5ed56acf3 for versions != 1.0.2 and in a later commit for versions
>= 1.1.0. Commit e3603397c434 introduced 3.x.x/conanfile.py where
these options were no longer deleted. This means the options were only
deleted for openssl >= 1.1.0 and openssl < 3.0.0 and there doesn't
seem to be a reason for it.